fix: a bar whose low is below zero decodes without taking the process down
A 31-bit price is sign-extended by subtracting 2^31, and that subtraction does
not fit the i32 it was done in. A spread quotes below zero, so the bar decoder
on the live path aborted on one.

Adds a suite that gives every wire parser what a socket can actually produce:
each prefix of a well-formed frame, that frame with a byte replaced at each
position, and runs that are not frames at all. The bar decoder was found by it
on its first run.

+++ src/control/historical.rs
@@ -724,8 +724,12 @@ pub fn decode_bar_payload(payload: &[u8], min_tick: f64) -> Option<crate::types:
 
     // Low price in ticks (31-bit signed)
     let low_ticks = read_bits(&mut pos, 31);
+    // Sign-extended in a width that holds the intermediate. A 31-bit value with
+    // its sign bit set is the raw value less 2^31, and that subtraction does
+    // not fit an i32 — a bar with a low below zero, which a spread has, took
+    // the process down.
     let low_ticks_signed = if low_ticks & (1 << 30) != 0 {
-        low_ticks as i32 - (1 << 31)
+        (low_ticks as i64 - (1i64 << 31)) as i32
     } else {
         low_ticks as i32
     };
